go/worker/storage/statesync: Move pruning to separate file
