+++ go/worker/storage/committee/worker.go
@@ -29,7 +29,6 @@ import (
 	storageApi "github.com/oasisprotocol/oasis-core/go/storage/api"
 	"github.com/oasisprotocol/oasis-core/go/storage/mkvs/checkpoint"
 	dbApi "github.com/oasisprotocol/oasis-core/go/storage/mkvs/db/api"
-	mkvsDB "github.com/oasisprotocol/oasis-core/go/storage/mkvs/db/api"
 	workerCommon "github.com/oasisprotocol/oasis-core/go/worker/common"
 	"github.com/oasisprotocol/oasis-core/go/worker/common/committee"
 	"github.com/oasisprotocol/oasis-core/go/worker/registration"
@@ -1360,44 +1359,4 @@ mainLoop:
 	// blockCh will be garbage-collected without being closed. It can potentially still contain
 	// some new blocks, but only as many as were already in-flight at the point when the main
 	// context was canceled.
-}
-
-type pruneHandler struct {
-	logger *logging.Logger
-	worker *Worker
-}
-
-func (p *pruneHandler) Prune(rounds []uint64) error {
-	// Make sure we never prune past what was synced.
-	lastSycnedRound, _, _ := p.worker.GetLastSynced()
-
-	for _, round := range rounds {
-		if round >= lastSycnedRound {
-			return fmt.Errorf("worker/storage: tried to prune past last synced round (last synced: %d)",
-				lastSycnedRound,
-			)
-		}
-
-		// TODO: Make sure we don't prune rounds that need to be checkpointed but haven't been yet.
-
-		p.logger.Debug("pruning storage for round", "round", round)
-
-		// Prune given block.
-		err := p.worker.localStorage.NodeDB().Prune(round)
-		switch err {
-		case nil:
-		case mkvsDB.ErrNotEarliest:
-			p.logger.Debug("skipping non-earliest round",
-				"round", round,
-			)
-			continue
-		default:
-			p.logger.Error("failed to prune block",
-				"err", err,
-			)
-			return err
-		}
-	}
-
-	return nil
 }
